requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
requests
gunicorn
Flask-Limiter
Flask-APScheduler
openai